backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.1
httpx==0.28.1
psycopg2-binary==2.9.9
qdrant-client==1.12.0
openai==1.57.0